+++ common/src/main/java/com/evandev/connectiblechains/CommonClass.java
@@ -3,6 +3,7 @@ package com.evandev.connectiblechains;
 import com.evandev.connectiblechains.config.ModConfig;
 import com.evandev.connectiblechains.entity.ModEntityTypes;
 import com.evandev.connectiblechains.networking.packet.ChainAttachS2CPacket;
+import com.evandev.connectiblechains.networking.packet.ChainSlackSyncS2CPacket;
 import com.evandev.connectiblechains.networking.packet.ConfigSyncPayload;
 import com.evandev.connectiblechains.platform.Services;
 import com.mojang.logging.LogUtils;
@@ -43,6 +44,12 @@ public class CommonClass {
                 ConfigSyncPayload.class,
                 ResourceLocation.fromNamespaceAndPath(MODID, "config_sync"),
                 ConfigSyncPayload::new
+        );
+
+        Services.NETWORK.registerClientReceiver(
+                ChainSlackSyncS2CPacket.class,
+                ResourceLocation.fromNamespaceAndPath(MODID, "s2c_chain_slack_packet_id"),
+                ChainSlackSyncS2CPacket::new
         );
     }
 }