+++ common/src/main/java/com/evandev/connectiblechains/client/render/entity/ChainKnotEntityRenderer.java
@@ -111,10 +111,9 @@ public class ChainKnotEntityRenderer extends EntityRenderer<ChainKnotEntity> {
         CatenaryRenderer renderer = getCatenaryRenderer(sourceItem);
 
         if (chainData.useBaked) {
-            ChainRenderer.BakeKey key = new ChainRenderer.BakeKey(startPos, endPos);
-            chainRenderer.renderBaked(renderer, vertexConsumer, matrices, key, chainVec, chainData.chainedEntityBlockLight, chainData.chainHolderBlockLight, chainData.chainedEntitySkyLight, chainData.chainHolderSkyLight);
+            chainRenderer.renderBaked(renderer, vertexConsumer, matrices, chainVec, chainData.slack, chainData.chainedEntityBlockLight, chainData.chainHolderBlockLight, chainData.chainedEntitySkyLight, chainData.chainHolderSkyLight);
         } else {
-            chainRenderer.render(renderer, vertexConsumer, matrices, chainVec, chainData.chainedEntityBlockLight, chainData.chainHolderBlockLight, chainData.chainedEntitySkyLight, chainData.chainHolderSkyLight);
+            chainRenderer.render(renderer, vertexConsumer, matrices, chainVec, chainData.slack, chainData.chainedEntityBlockLight, chainData.chainHolderBlockLight, chainData.chainedEntitySkyLight, chainData.chainHolderSkyLight);
         }
 
         matrices.popPose();
@@ -167,6 +166,7 @@ public class ChainKnotEntityRenderer extends EntityRenderer<ChainKnotEntity> {
             renderChainData.chainHolderSkyLight = level.getBrightness(LightLayer.SKY, blockPosOfEnd);
             renderChainData.sourceItem = chainData.sourceItem;
             renderChainData.useBaked = chainHolder instanceof HangingEntity;
+            renderChainData.slack = chainData.getSlack();
             result.add(renderChainData);
         }
         state.chainDataSet = result;