+++ common/src/main/java/com/evandev/connectiblechains/networking/packet/ConfigSyncPayload.java
@@ -2,7 +2,7 @@ package com.evandev.connectiblechains.networking.packet;
 
 import com.evandev.connectiblechains.CommonClass;
 import com.evandev.connectiblechains.client.ClientInitializer;
-import com.evandev.connectiblechains.util.Helper;
+import com.evandev.connectiblechains.util.MathHelper;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
@@ -11,7 +11,7 @@ import org.jetbrains.annotations.NotNull;
 public record ConfigSyncPayload(float chainHangAmount, int maxChainRange,
                                 boolean collisionsEnabled) implements CustomPacketPayload {
 
-    public static final Type<ConfigSyncPayload> TYPE = new Type<>(Helper.identifier("s2c_config_sync_packet_id"));
+    public static final Type<ConfigSyncPayload> TYPE = new Type<>(MathHelper.identifier("s2c_config_sync_packet_id"));
 
     public static final StreamCodec<RegistryFriendlyByteBuf, ConfigSyncPayload> STREAM_CODEC = StreamCodec.of(
             (buf, payload) -> payload.write(buf),